Time: 1 ms (99.98%), Space: 42.7 MB (26.72%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.java
@@ -12,7 +12,7 @@
            rev = rev * 10 + digit;
         }
 
-        return rev;
+        return (int)rev;
     }
 
 }
